refactor application configuration files
