+++ src/main/java/com/omega/jobportal/JobPortalApplication.java
@@ -2,10 +2,13 @@ package com.omega.jobportal;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.scheduling.annotation.EnableAsync;
 
-@SpringBootApplication
+
 @EnableAsync(proxyTargetClass = true)
+@EnableJpaAuditing
+@SpringBootApplication
 public class JobPortalApplication {
 
     public static void main(String[] args) {
