+++ src/main/java/com/omega/jobportal/config/SecurityConfig.java
@@ -3,7 +3,7 @@ package com.omega.jobportal.config;
 import com.omega.jobportal.auth.CustomOauth2UserService;
 import com.omega.jobportal.auth.GlobalAuthenticationEntryPoint;
 import com.omega.jobportal.auth.Oauth2LoginSuccessHandler;
-import com.omega.jobportal.user.UserService;
+import com.omega.jobportal.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -13,6 +13,9 @@ import org.springframework.security.authentication.dao.DaoAuthenticationProvider
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler;
@@ -23,11 +26,10 @@ import org.springframework.security.web.context.SecurityContextRepository;
 @RequiredArgsConstructor
 @EnableWebSecurity
 public class SecurityConfig {
-    private final UserService userService;
-    private final PasswordEncoder passwordEncoder;
     private final GlobalAuthenticationEntryPoint globalAuthenticationEntryPoint;
     private final Oauth2LoginSuccessHandler oauth2LoginSuccessHandler;
     private final CustomOauth2UserService customOauth2UserService;
+    private final UserRepository userRepository;
 
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
@@ -50,8 +52,8 @@ public class SecurityConfig {
 
     @Bean
     public AuthenticationManager authenticationManager() {
-        var authenticationProvider = new DaoAuthenticationProvider(passwordEncoder);
-        authenticationProvider.setUserDetailsService(userService);
+        var authenticationProvider = new DaoAuthenticationProvider(passwordEncoder());
+        authenticationProvider.setUserDetailsService(userDetailsService());
         return new ProviderManager(authenticationProvider);
     }
 
@@ -64,4 +66,17 @@ public class SecurityConfig {
     public SecurityContextLogoutHandler securityContextLogoutHandler() {
         return new SecurityContextLogoutHandler();
     }
+
+    @Bean
+    public UserDetailsService userDetailsService() {
+        return username -> userRepository.findUserByEmail(username)
+                .map(SecurityUser::new)
+                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
+    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
 }
